+++ src/store/index.js
@@ -4,7 +4,7 @@ import permission from './modules/permission';
 import monitorinstall from './modules/monitoringInstall';
 import user from './modules/user';
 import account from './modules/account';
-// import roles from './modules/roles';
+import roles from './modules/roles';
 import getters from './getters';
 
 Vue.use(Vuex);
@@ -14,7 +14,8 @@ const store = new Vuex.Store({
         permission,
         monitorinstall,
         user,
-        account
+        account,
+        roles
     },
     getters
 });
